BE/pyproject.toml
[project]
name = "021-hackx-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beanie>=2.0.1",
    "crypto>=1.4.1",
    "cryptography>=46.0.3",
    "dotenv>=0.9.9",
    "email-validator>=2.3.0",
    "fastapi>=0.124.4",
    "motor>=3.7.1",
    "passlib[argon2]>=1.7.4",
    "pycryptodome>=3.23.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.5",
    "uvicorn>=0.38.0",
]

